CUSTool/requirements.txt
pynput>=1.7.6
watchdog>=3.0.0
pytest>=7.0.0
mock>=4.0.3
pygetwindow>=0.0.9
pyautogui>=0.9.54
Pillow>=9.0.0
psutil>=5.9.0
pywin32>=304
